requirements.txt
Flask==2.0.1
pymongo==3.12.0
python-dotenv==0.19.0
Flask-Login==0.5.0
Werkzeug==2.0.1
email-validator==1.1.3 